pyproject.toml
[tool.black]
enable-unstable-feature = [
    "hex_codes_in_unicode_sequences",
    "string_processing",
    "hug_parens_with_braces_and_square_brackets",
    "unify_docstring_detection",
    "no_normalize_fmt_skip_whitespace",
    "wrap_long_dict_values_in_parens",
    "multiline_string_handling",
    "typed_params_trailing_comma",
]
fast = true
line-length = 119
preview = true
required-version = "24.8.0"

[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "alarm-clock"
packages = [{include = "alarmclock", from = "src"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
flake8 = "^7.1.1"
flake8-bugbear = "^24.8.19"
flake8-isort = "^6.1.1"
pep8-naming = "^0.14.1"
playsound3 = "^2.2.1"
python = "^3.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pre-commit = "^3.8.0"

[tool.isort]
include_trailing_comma = true
line_length = 119
multi_line_output = 3
profile = "black"
src_paths = ["src", "tests"]
use_parentheses = true

[build-system]
requires = ["poetry-core", "setuptools", "wheel"]
build-backend = ["poetry.core.masonry.api", "setuptools.build_meta"]
